Change lvs2014 IP on private1-a3-codfw vlan
It seems there was a mix-up with the manual IP allocation on this
vlan and 10.192.5.5 was assigned to mw2379 also.

As there are no backends on that vlan right now it's easiest to
change the lvs IP than the mw server so doing so via this
patch.  Netbox already updated.

Change-Id: I8fdd87ca0913403da3c8495f15ffcc9b3472b48b