etherpad: make exporter and blackbox checks configurable
This change makes the prometheus-etherpad-exporter listen IP and port
configurable. Also the exporter and blackbox checks depend on the
service_ensure flag now.

The Deb package default listen address ":9198" does not work on bookworm.
With this address the exporter fails with
socket.gaierror: [Errno -2] Name or service not known
So facts['networking']['ip'] is used as the listen IP for the exporter.

Bug: T316421
Change-Id: Ib064b12e16f19122fff336611b4f62500111981c