haproxy: configure extended logging (preparatory for Benthos)
This commit introduces the `profile::cache::haproxy::extended_logging`
hiera conf that allows to switch from the current log-format to a more
rich one that Benthos will need to perform it's processing.

With this extended logging more request/response headers needs to be
captured.

Extended logging should not break the current mtail configuration but
a small modification in that regard has been made also to be extra
sure.

The `newlog` placeholder isn't actually needed but can be useful in
this first step to differentiate between the old log format (used by
mtail) and the new one (used by Benthos). When Benthos will be
actually deployed on cp hosts and mtail dismissed, the old (first)
part of the log-format will be deleted.

Bug: T358105
Bug: T358107
Change-Id: I5c52c40db2eaab3049741411d98a3d48345581a0