ferm: Check ferm.service status in ferm_status.py
The current ferm_status.py script, used by puppet to check if it needs
to restart the ferm service, does not check for a bad systemctl status
ferm.service.
This means that even when ferm is in a bad state, puppet does not
restart it.
Adding `systemtctl is-active --quiet ferm.service` to ferm_status.py and
exiting early in case of a non-zero return should fix this.

Bug: T354855
Change-Id: I891c2bcec9aad68d0a06ac510f7ff1018f399224