cache: start using benthos on single host for haproxy log parsing
This commit introduces the ability to install benthos for cache hosts
to process HAProxy log files. The actual installation is protected by
the profile::cache::base::use_benthos hiera switch that will default
to true when properly tested.

As for other hosts with the ::profile::benthos, the actual instance(s)
configuration is defined in hiera.

Bug: T358109
Change-Id: Ifa7d333042b30fabd6767cbf770df81da22e67ce