prometheus::ops: monitor active etherpad instance only
In T316421 a new etherpad host was introduced. The old host and
a failover host in codfw remain online but with disabled etherpad
service.
The disabled hosts cause JobUnavailable alerts due to the prometheus::class_config
pointing to a disabled exporter. Also running the exporter with disabled
etherpad creates ExporterUnavailable alerts. Running multiple etherpad
instances on the active and old hosts is not possible due to consistency issues.

This change tries to fix this issue by setting the class_parameters
service_ensure for the etherpad exporter.

Bug: T316421
Change-Id: I066a0eac72fe188cef87001a147053b6289915c9