services_proxy: Add mw-parsoid in the mesh
Why:
The previous approach, based on LVS failed, due to scap utilizing the
same data structure to deploy MediaWiki. So we fallback to utilizing the
service mesh for the migration to mw-on-k8s.

What:
Add mw-parsoid to services_proxy stanzas
Enable it for RESTBase

In subsequent puppet patches, we will gradually move RESTBase to it

Hosts: auto
Bug: T357392
Change-Id: I951e7f374ffdcfb2dc7cb357b706f0aa20e1d478