profile::base: Allow running without cron installed
If enabled, don't depend on the presence of a cron daemon. In a
standard installation we still have common packages which depend on a
cron-compatible daemon, but there are already use cases in Cloud VPS
where cron isn't necessary. With increased adoption of systemd timers,
this might also be applicable for a future baremetal installation. For
now this option only omits the automated service restarts for cron.

Bug: T358343
Hosts: idp1002.wikimedia.org, mw1399.eqiad.wmnet, idm1001.wikimedia.org
Change-Id: I254a7ef74efbe5e960a83162c015f593438ada71