P:dns::auth::update: move authdns-update state to confd
dns6001 has management of authdns-update state via confd and we are
ready to roll this out to all other hosts as well. This means that if
you mark a DNS host depooled for the authdns-update service from
confctl, it will be removed from all hosts for the authdns-update.conf
file, which we generate via confd.

Bug: T347054
Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: Ief809592478fdc0309a0101b05c2f297654d7eaa