phabricator: add scap::user setup to migration profile
The "migration" role/profile is to be applied to new or freshly
reimaged phorge servers.

It allows for syncing data and sets up scap before the first
deployment of the application.

Adding the scap::user here is supposed to prevent the bootstrap
issue that happened last time we reimaged the production server.

Bug: T357572
Change-Id: I64a151a88653cd24dd0db2a0dfd763d206a2697f