cache::text: remove git.wikimedia.org
https://git.wikimedia.org redirects to Phabricator.

This is confusing since it's none of the repos we actually use such as
https://gerrit.wikimedia.org and https://gitlab.wikimedia.org and repos
on Diffusion have been phased out.

There was an original suggestion to redirect this to a landing page that
explains this but further commenters said they prefer to drop it
entirely since users have already been used to disregarding it.

So this goes the simple route and just removes it from the caching
server layer, which is still easy to revert, before doing other cleanup.

Bug: T323073
Change-Id: I52dbf418cf6e6d5ec0eeb0f5e2025a4416fdc5b4