phabricator: remove git.wikimedia.org vhost, rewrites and tests
https://git.wikimedia.org redirects to Phabricator.

This is confusing since it's none of the repos we actually use such as
https://gerrit.wikimedia.org and https://gitlab.wikimedia.org

There was an original suggestion to redirect this to a landing page that
explains this but further commenters said they prefer to drop it
entirely since users have already been used to disregarding it.

This removes the existing vhost with the special rewrite rules the
"gitblit host" parameter and the related tests entirely after
I52dbf418cf6e6d5e previously deactivated it on the caching server layer.

Bug: T323073
Change-Id: Id624fdf1f13991106affc090a07ed6b751d1eee1