admin: Remove the os-installers group
Previously, this was for members of sre-admins (new SREs without global
root yet) to be able to run wmf-auto-reimage scripts, but those scripts
haven't existed since 2021.

Instead, add members of sre-admins to datacenter-ops, which has access
to `secure-cookbook sre.hosts.reimage` for the same purpose. (It has a
bit more access too, but we're comfortable with that.)

The reimage cookbook still requires access to the mgmt password, which
sre-admins members *don't* have (since they don't have access to pwstore
yet) but there's a workaround discussed on the linked Phab task, which
we'll incorporate separately into training materials.

Bug: T358361
Change-Id: Ife3e909a14ed485a9b12102f1c738ff2727c1f60