mediawiki: Restrict /wiki RewriteRule
Currently this matches /wiki/Foo (as intended) but also /wikifoo
(accidentally). That produces a weird behavior where /foo is a 404 but
/wikifoo is a 301 to the vhost's main page.

It originates as an artifact from when this RewriteRule directive used
to be a ProxyPass, and before that an Alias, both of which act on paths
rather than regular expressions.

Instead, only match /wiki/Foo (or /wiki/Foo/Bar, ...). The special cases
/wiki and /wiki/ aren't well-defined, but a 301 to the main page is as
reasonable behavior as any, so keep it as-is.

Bug: T357595
Change-Id: I61e09b49f465b96b76a7cb96f3e35428ac1fc312