Update analytics sqoop jobs
This patch splits the refinery-sqoop-mediawiki-production script
in two: refinery-sqoop-mediawiki-production-history for tables needed
by mediawiki-history, and refinery-sqoop-mediawiki-production-not-history
for tables not needed by mediawiki-history. This replicates the pattern
we already use t sqoop from labs.
The new scripts are then used in refinery-sqoop-whole-mediawiki instead
of the old one, with an order allowing to get data for mediawiki-history
faster.

Change-Id: I83cd03ddeadb552262ca3f849f332de9cea95bb3