Add DSCP marking options to current firewall classes
Adds a 'qos' parameter to firewall::service and firewall::client
that allows teams to specify a DSCP marking to set for traffic
matching the rules.

Additionally provides a generic nftables::rules to allow for more
complex rules to be manually specified, which is needed for some
edge-cases.  This could be done less-generically to fit certain
use-cases, but I think we will need the flexibility so kept it
simple.

Lastly the base nftables and ferm rules are also extended to mark
any traffic not matched by specific rules as DE/CS0, placing it in
our 'normal' priority class across the  network.

Bug: T339850
Change-Id: I1095c534d3000b576d057b9d68d70f02627e111f