Change the default systemd timer email source to noreply@wikimedia.org
We have decided to change the default source email address for the
systemd-timer emails so that it no longer uses a FQDN. The reason for
this is to increase compatability with destination mail servers that do
not like unroutable domains (e.g. *.wmnet).

As well as changing the default address, we are adding a new --mail-from
option to the systemd-timer-mail-wrapper.py scripts. This is so that if
we wish to do so we can override this behaviour and configure a custom
address that will be used for both the From: and Sender: headers.

Bug: T358675
Change-Id: Ice17ef4fd321c2a03f1a63d0c334c8c92ada8dcc