Allow systemd::timer::job to send from a custom address
We wish to allow systemd timer jobs to be able to use a custom Sender
and From headers, in order to avoid using the default FQDN values that
are unroutable and rejected by some destination mail systems.

We have added the new '--mail-from' option to the
systemd-timer-mail-wrapper.py script in another commit. This commit adds
the option to add a custom address to the systemd::timer::job defined
type, but it is intended to fail safely to the existing behaviour.

Bug: T358675
Change-Id: I16f7db5670ac6c6debe58190d0b90ee639bb446c