P:dns::auth: add support for depooling services via confd/confctl
In the series of previous commits under T347054, we are managing the
state of the DNS hosts dynamically via confd for dns6001. One of the
most important parts of this is to pool or a depool a host via confctl
by controlling the bird advertisements. This commit does that by
creating a file that has the current status of a host (pooled or not)
for each type of service (ntp, recdns, authdns); a healthcheck wrapper
script then causes the healthcheck to pass or fail, which then withdraws
the route from anycast-healthchecker and then eventually bird.

We do this for all type services on the authdns hosts. Additionally, we
also update the service_type names for all services and make a
distinction between authdns and authdns-nsX, matching the defined state
in conftool-data and the etcd layout.

Bug: T347054
Hosts: dns1004.wikimedia.org, dns6001.wikimedia.org
Change-Id: I28c395e9067fbd4f35fd1c0706433cc17153a55f