Move lvs2012 from private1-b-codfw to private1-b2-codfw vlan
This patch needs to be merged when we reimage lvs2012 and change its
vlan from private1-b-codfw to private1-b2-codfw.

The change adjusts the BGP peer the host will connect to from the
core routers to lsw1-b2-codfw, and adds a new vlan sub-interface to
the primary port so it can still connect to realservers on the old
vlan.

Bug: T352918
Change-Id: I1ed8a98f2a6bfb984d5a5203f043a43ef80c7e8c