Move lvs2011 from private1-a-codfw to private1-a2-codfw vlan
This patch needs to be merged when we reimage lvs2011 and change its
vlan from private1-a-codfw to private1-a2-codfw.

The change adjusts the BGP peer the host will connect to from the
core routers to lsw1-b2-codfw, and adds a new vlan sub-interface to
the primary port so it can still connect to realservers on the old
vlan.

Bug: T352920
Change-Id: I1e45374e74bc9001f0da2681fc665ed21bb9649e