P:openstack: rabbitmq: restrict clustering ports
Only allow hosts specifically listed in rabbitmq_nodes or
rabbitmq_setup_nodes to access the clustering ports. This is done by
limiting the current cloud-private term to only contain the access
ports. The type hint of $cloud_private_supernet is also fixed to not be
optional.

Also, drop an outdated comment. There are no more terms that allow
access for hosts that previously used the cloud-hosts interface but now
use cloud-private.

Change-Id: I21a46bee0e90db10ffeb2e503be434ab3a1de8b6