openstack: neutron: manage ml2 plugins directory
Instead of removing specific files we do not want, remove everything in
that directory not provisioned by Puppet. This way when we introduce
code to manage the OVS agent we can just drop that into the directory
withoutu changing the common code and it'll work fine.

According to Cumin we have no unmanaged files that this would remove,
excluding an empty placeholder file[0].

[0]: https://phabricator.wikimedia.org/P58286

Bug: T326373
Change-Id: I4cbf8dfda3dbc1c7cdc20a3cfb2fd4d6f26de7bc