dns::auth: move all service state management to confd
In the previous series of commits under T347054, we have progressively
added state management of the DNS hosts to confd for the dns6001 host so
that we could safely test those changes. This commit rolls those changes
out to all DNS hosts.

We remove the confd_enabled boolean and update the hieradata for DNS
boxes, completing the migration to confd for the state/service
management of all DNS hosts.

Bug: T347054
Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org, dns2004.wikimedia.org
Change-Id: Idd7c52bdd0b27a36cec6e3d9166468e619459709