ci_test: add profile::ci::website to allow deployments
The profile::ci::website profile includes the apache site
for integration.wikimedia.org but also the scap::target
to allow scap deployments for it.

scap::target pulls in scap::ferm and that pulls in
firewall::service deployment-ssh allowing deployment
servers to ssh to scap targets.

Regardless of the integration.wikimedia.org this is needed
to allow any scap deployments to the hosts.

In this case we want to deploy zuul but the way puppet is setup
doesn't make this obvious. It's not pulled in from the zuul profiles.

And with the missing firewall rules any attempt to deploy zuul
simply times out.

By default only bastion hosts are allowed to ssh to any other hosts
so it can be confusing for users that they can talk to the machine
from a bastion but not from a deployment host as it was reported to us.

Bug: T358237
Change-Id: I26dc6ea7d692ccc2085f73408901b72f198b3fd5