eqiad1::pdns: remove check_dns alerts
This dns service isn't listening on the internal IP of the DNS servers
so these tests always fail. The dig tests that immediately follow check
the proper server IPs so should be adequate tests.

Ideally the check_dns logic would support passing in an alternative
server IP but since icinga is being phased out I'm not going to mess
with it.

Change-Id: Id67b64cb4247e899fb9e06981727e12046f62c94