P:dns::auth: skipping running authdns-update on host if not pooled
This commit is based on a question from Riccardo and a suggestion from
him: if a given DNS host is not pooled in confd, we are still running
authdns-local-update on it if authdns-update was run from that host.
That might not be the best behaviour however as most people SSH to the
DNS host in their backlog or from their muscle memory (most of the time
dns1004 or some host in eqiad) and if dns1004 was depooled in confd, we
should not be pulling in updates to it.

This commit checks if the host is in the pooled list and if yes, it will
run authdns-update on itself. If not, it will skip it over.

Bug: T347054
Hosts: dns6001.wikimedia.org, dns1004.wikimedia.org
Change-Id: I64ad7e02c7ab8c138d6183f7006a14dd6af6e361