P:prometheus::ops Remove new LDAP hosts from Prometheus
Our new Bookworm LDAP servers do not have the LDAP exporter that the
Buster hosts have, as this has been removed from Debian. Prometheus
reports the exporters as down, triggering an alert.

Change the class to trigger Prometheus target inclusion on to the class
for the exporter, not the LDAP role.

Bug: T359524
Change-Id: Ia11bf48d2de4556ac08fadec16a8b5429bfc423b