Remove Tomcat spec tests
This was once added for an older version of Tomcat, but it's become a nuisance
to keep it in sync with what we're actually implementing on the application
side: We have two different Tomcat 9 builds and also Tomcat 10 in our
WIP CAS setup. We're testing all functionality on idp-test anyway.

Bug: T359333
Change-Id: I4649d90b484a9fb291a294dbbe5742761a314a3b