prometheus: ethtool_exporter: use package{} directly
Using ensure_packages() messes up with Puppet's implicit ordering
resulting in the first Puppet run trying to start the service beforoe
the package has been installed. Given that realistically no other Puppet
class will try to install that package this patch turns it to a direct
package{} resource. The other option would be to add an explicit
dependency to the systemd::service resource but I think this is a nicer
solution.

Change-Id: I219de94afe6894fb678e9d2529b06c113c9d5c89