mw-xml.sh: Update maintenance script
In T359643 I want to get rid of the /srv/mediawiki/php symlink, which
means eradicating all uses of it.

Previously mw-xml.sh would cd to /srv/mediawiki/php at the start of
the script, perform the dump, and then cd back to wherever it started
at the end of the script.  mw-xml.sh and maintenance/dumpBackup.php do
not depend on the current directory being /srv/mediawiki/php, so I
removed the change of directory.

Tested by running
```
/usr/local/bin/mwscript maintenance/dumpBackup.php \
   labswiki  --current --uploads \
   | sed 's/<model>yaml<\/model>/<model>wikitext<\/model>/g' \
   | sed 's/<format>application\/yaml<\/format>/<format>text\/x-wiki<\/format>/g'\
   | head
```
from my home directory on deploy2002.

Bug: T359643
Change-Id: I8b3b09c5de83265ad3fcd8289d9de8acfbbbca10