profile::thanos: Add latency histogram buckets back for Istio
With the setup introduced in patch 989458, we started having a recording
rule that covered 20 latency buckets (all that Istio exports).

This patch roughly halves the number of buckets recorded, which should
help with Thanos not being fast enough to evaluate the timeseries into
these recording rules.

It also drops the unneeded mention of `le` in the recording rule for
istio_sli_latency_request_duration_milliseconds_count:increase5m, since
the `..._count` metrics do not have that label anyway, so this was a
no-op, but misleading.

Bug: T359879
Change-Id: Icc408258e376a652b5958fa167e530a1fc17d463