profile::prometheus::k8s: move istio metrics to a separate job
We have been trying to reduce the time series ingested from Istio
for a while, but so far we have bumped into some blockers:
- Reducing Istio metrics at the source seems not working as expected,
  at least with our version of Istio.
- Dropping labels from Prometheus is too risky with the current config
  since we use a single job for all pods, and the Istio labels may
  have generic names (so dropping them could have unwanted side-effects
  to other time series).
The idea of the patch is to exclude the istio metrics from
the k8s-pods job (they all carry an annotation like
sidecar.istio.io/inject=true|false) and to collect them in a separate
job, in which we can drop labels freely without risking side effects.
This change shouldn't break any current dashboard since we don't
filter anywhere for job.

Bug: T351390
Change-Id: Ib02be01715bf8af62bc6cef177893d4ce532d5f6