P:cumin: add alias for dnsbox hosts (dns-rec/auth)
We already have aliases for dns-rec and dns-auth that basically mean
dnsbox since the rec and auth roles were merged in T330670. There is
value in keeping dns-rec and dns-auth but we should have a unified
dnsbox alias that we will then use for the restart cookbook.

All of these mean the same thing and ideally we should remove dns-rec
and dns-auth aliases and move everything to dnsbox but that cleanup is
for another commit and once everything has moved to the new confd
configuration that we did in T347054.

Change-Id: Id881f31adb136b29c6db97263b3e1a9cc45640ca