gitlab: remove duplicate key type from gitlab known_hosts
The known_hosts entry for Gitlab has a duplicate key algorithm which
leads to problems when pushing to GitLab. The content of the GitLab
public key contains the algorithm (ssh-ed25519) already and @@sshkey
appends another type.
So this change tries to just remove the dedicated key type in @@sshkey
and just use one already included in the key file.

Bug: T337107
Change-Id: If0e9d14bd503661f0ab4a80a01391aa7ed3c8e60