Also disable monitoring for AQS1
Disable profile::aqs::monitoring_enabled, which only controls the presence
of the (now failing) health check for the AQS root url.

Also pass the parameter down to the monitoring::service define (previously
it was only added, but had no method to also remove it).

Also yank the not obsolete TODO item.

Bug: T360522
Change-Id: I2006ef1cbd4634309368be1050673018737bb67e