role::aqs: deploy the PKI-enabled TLS bundle and use it on aqs1010
This change rolls out a special TLS truststore to aqs1010 that contains
two Root CA certs:
- PKI root
- Cassandra CA manager root
The idea is to have a new truststore that contains the PKI CA Root cert,
so that in the future the node will be able to connect to other
Cassandra instances using PKI certificates (basically we'll be able
to upgrade one node at the time).
The change is applied only to aqs1010 as starting test, it will
be hopefully applied to all other nodes as well.

Bug: T352647
Change-Id: I61d4b312e77ed99132595ea8cd4ca9c6fd5ad7f0