[gitlab] Move backup script locking out of main script root
Currently we place a lockfile on any run of the script, however some
operations (notably the help function, and the unlock function) won't
run if another instance is running.

The other option is to check `backup_type` before we place a lockfile,
but this seems like a neater/cleaner option, even if it's a little more
verbose.

Bug: T358559
Change-Id: I8a2df9cee45f2ac7a372d8336d6d8ff05337c5d1