Update the from address of all email from refinery jobs.
We wish to avoid using non-routable addresses in the from addresses of
all notification emails in future. The default from address for refinery
currently uses the fully qualified domain name, which is non-routable.

This commit updates the default settings for all systemd based refinery
jobs, such that the emails originate from noreply@wikimedia.org.

We will also update the default in refinery-source and review how email
alerting works as part of any migration to airflow.

Bug: T358675
Change-Id: I4daca805fb044ce913a3f081653b645212430d90