prometheus/ops: add config for scraping apache metrics on planet servers
As the TODO on line 753 says in the future this should ideally be
generated for all hosts using an apache.

But for now it requires a special config snippet for each role/class
to tell prometheus to scrape apache metrics.

The exporter has recently been fixed to work on bookworm hosts with
I81c1f8e7c65c53b and could then be added to planet servers.

Now adding the special config for it.

Bug: T359556
Change-Id: I83b412c855dee7fb717987838bf18b7036619797