P:pybal: add a check to ensure Pybal service has been restarted
As has happened many times in the past and today, it's possible that the
Pybal configuration file was changed but the Pybal service was not
restarted immediately after. Puppet does not handle the restart of Pybal
so it is on us to do that.

What this results in is that a change is made and we forget to restart
Pybal. This means that a few days or weeks later when Pybal is restarted
for maintenance work, it reads the configuration file with a change that
was not applied since the file was changed.

To prevent this from happening and as used in the case of the NTP
service and elsewhere, we should alert in case the configuration file
for Pybal was changed but the service was not restarted.

Hosts: P:lvs
Change-Id: I92857ce7db26522ab78863535d566526cec541e9