P:puppetserver::git: do not mark directories as safe
The safe.directory config was deployed to fix a permission issue during
initial provisioning (https://phabricator.wikimedia.org/P58697).
However, on a closer look, the fix is wrong: both the `git clone` exec
and `git remote set-url` should be running as 'gitpuppet' so the
permissions should be correct.

Furthermore I'm seeing a few instances that are failing to run
puppet-git-sync-upstream.service due to permission errors. So by
removing the safe.directory we should prevent anything from writing in
the git directory as root (as everything should happen as gitpuppet).

Change-Id: Icef9df4df3555e2b18b3978b8c63475822f32308