scap: Add retry_on_timeout to scap httpbb checks
Why:
The first try of parsing expensive pages, like Barack Obama on enwiki,
is usually slow, and doesn't indicate a regression. It has the unwanted
side-effect of causing confusion for deployers when the tests are
running during a deployment.

What:
Adding --retry_on_timeout to the scap httpbb test will retry long pages
a second time, and is consistent with our regularly run httpbb
functional tests. This should alleviate this false positive for
deployers.

Bug: T360867
Change-Id: I364b2b7b00570d3f7bafdf6ff2fe78e20d281cb6