[gitlab] Narrow scope of gitlab backup rsync commands
If rsync runs while a backup is being created on a replica (either for a
failover, or for a test), then these files will be deleted, interrupting
the operation.

If we narrow the scope of the rsync to only transfer the completed
backup files (always either named with *_gitlab_backup.tar or
gitlab_config_*.tar), this should not interfere with other operations,
and the pre-existing lock files can help with the rest.

Bug: T361219
Change-Id: I965cd9bb7c54e982a80567c96d48e47cdb2b995b