profile::wmcs::kubeadm::etcd: install etcd package before referencing uid
Previously this demanded that these files be created before
the package was created; that doesn't make sense since the etcd
user wouldn't exist before the package. I'm replacing that with
a bump of the etcd service which will hopefully fix whatever race
we were trying to fix with the broken ordering.

Hosts: toolsbeta-test-k8s-etcd-21.toolsbeta.eqiad1.wikimedia.cloud
Bug: T349207
Change-Id: If4095d067ccc0f5ee095eaa965afd56a9c420c4d