toolforge: Exclude /usr/bin/sudo from Wheel of Misfortune
Add /usr/bin/sudo to the excluded process list in
wmcs_wheel_of_misfortune.py so that it will not attempt to kill `become
$TOOL` interactive sessions. This does introduce a WP:BEANS possibility,
but I think we can worry about how to deal with that if we find it being
abused in the wild.

Patch inspired by receiving a "sudo killed by Wheel of Misfortune on
tools bastion" email from the script that terminated my `become
wikibugs` interactive shell.

Change-Id: Ifda04cdb318dba3aeecb4e88338daf36b0cfaf11