httpbb: add missing virtual hosts to legacy miscweb tests
I am touching the SSL certificates on these machines and
would like to have tests that cover all the virtual hosts
that are actually still served from here.

The experimental sites and commons-query are configured as apache
sites but were not in the tests.

I wanted to prevent that we miss any of them on in the SAN
field of the cert, breaking those sites while doing I146bd03ab24e6040.

Therefore I want a test that works before and after.

Bug: T360413
Change-Id: Ib5e1cdab652a1df985612098940a9dfbe89d2703