WMCS: Fix type of ipb_range_start and ipb_range_end in the b/c view
ipb_range_start and ipb_range_end are non-nullable in the old schema, so
if they are null in the new schema, convert null to an empty string.

I left it as nullable in the ipblocks_ipindex view, since if you're
using that view, you explicitly want conditions on IP address fields
to use indexes, and coalesce() would probably break that.

Bug: T355034
Change-Id: I39c3c41f9647ec655916d92c653e4618b70458db