benthos: add metric for ttfb
First test to add metric for ttfb.

The logic has been enclosed in a switch processor to mimic the
current mtail behavior that increments conditionally the two counters
just if ttfb > 0 and the healthcheck one only if the request it's
actually an healthcheck.

The "delete_unused_field" processing has been moved to the very end to
leverage intermediate parsing results for metrics before they are
dropped.

Bug: T361845
Change-Id: I5ace61514c3a4e32edb1b1ac3499a5bef9a253ec