os-reports: Update puppetdb query
Previously we were matching for the system::role define which was
historically added to every role. However, recently support was added
to make this step obsolete and simply declare the description via
the profile::base::production::role_description Hiera parameter to
the base profile.

Update the puppetdb query to match for the role title instead.

Bug: T355924
Change-Id: I9119bd3463b25cd41f3ecbeac583591293df1cd3