profile::prometheus::analytics: remove old metric relabeling
In T193017, I had to workaround some metric names
to consolidate Cassandra dashboards. At the moment this seems
to be the only big difference with the services' Prometheus
instance, and the main visibile effect is that metrics like:

cassandra_table_readlatency{table="peers",keyspace="system",}

are exposed via the prometheus exporter but mangled by
Prometheus when ingesting (namely dropping the table label).

Change-Id: I1143489ab547e828fc0d03350e6b1bec8ce40fba