gitlab_runner: temporary allow dockerfile frontend on gitlab-runner2004
To self-build the dockerfile frontend image (which is also based on
a Dockerfile) we have to allow dockerfile frontend on Trusted Runners
temporarily. Once the image is published successfully on our
registry we can revert that change.
In the last try I9772e6d5846b3ae9d6ad3f0e70436d9338571d76 a wrong setting
was used to self-build the dockerfile frontend image.

Bug: T357612
Change-Id: I000f47d771f5437c07ff8ab0a04a89bc79e39cc0