prometheus: add aggregate metrics for benthos
Similarly to what's already present for mtail-generated metrics above.

The aggregated metric for mtail keeps the same name but are built
specifying explicitly the job label.

The benthos aggregate metrics has different name to easily
differentiate, will probably be discarded when the transition to
Benthos will be completed and the former aggregated metrics will be
built with job=~"benthos".

Bug: T361845
Change-Id: I37caa14311d01a012ac9a12d2646482de0949c64