cassandra: move cqlshrc's template to wmf-ca-certificates.crt
The file wmf-ca-certificates.crt is deployed on all production
nodes, and it can be tweaked to include other certs in the bundle.
If the tls_use_pki_truststore option is enabled it means that
all Cassandra instances are using it, and they are potentially
going to run PKI TLS certs as well. Update cqlsh accordingly
to trust the .ctr composed by the right CA certs.

Bug: T352647
Change-Id: I0b068eb52047b0f443e1f977abf81bf8aa4da419