cassandra::instance: fix PKI keystore for each instance
In the current code the Cassandra instance's keystore resource
is called "cassandra_keystore", that works when we have only
one instance for each host (like ml-cache, the first cluster
running PKI) but that leads to duplicate resource declaration
for 2+ instances on the same node.
Add the hostname to the resource definition to avoid name
clashes.

Bug: T352647
Change-Id: I78f18f5e6c3392c9f0163d98fccb530c64ed107d