Create a new and profile for the new matomo server
We intend to replace matomo1002 with matomo1003, upgrading matomo itself
from 3.14.1 to 4.16.1 and the host from buster to bookworm.

Matomo was previously named Piwik, so we still apply a 'piwik' profile
to matomo1002, although this profile loads a 'matomo' module.

In order to facilitate setting up the new host without adversely
affecting the production service, we are going to duplicate the piwik
profile to matomo.

Once the service has been migrated to the new host, we will remove the
unused piwik profile.

Hosts: matomo1002.eqiad.wmnet,matomo1003.eqiad.wmnet
Bug: T351552
Change-Id: I2b586460511528769c84eb64556aaf7b23950b35