profile::wmcs::backup_cinder_volumes: remove remove-unhandled-backups timer
In case of a coding error (or a switch to new hardware) this would delete
all of our backups and require a fresh start. We already have code
to remove expired backups which in normal usage should clean up these
unhandled backups as time passes; no need to wipe them out immediately.

Change-Id: I85a32ed4beb637d51accead597bfc94b6248ae9d