graphite: avoid including multiple roles, define primary host in hiera
Avoid a style guide violation, including multiple roles on a node
and having to edit site.pp to change the primary server.

Move definition of the primary host to hiera/common.yaml alongside
other definitions of primary or active servers.

Delete role that isn't an actual server role and include both
alerting profiles in the production role.

Still ensure alerting only happens on the active server and allow
for switching it with just a Hiera edit.

Change-Id: Ifbc085e3b6cdc2318a6f6168845d1ca0f041f50d