memcached/mcrouter: remove onhost memcached
What?
We are removing anything related to onhost memcached

Why?
Onhost memcached was an ambitious attempt to improve performance
by running a small local memcached instance within an appserver.

In the longrun, for our workload, the performance gain was slightly
visible under *really heavy loads*, thus not worth the effort
to keep it around.

Bug: T345740
Change-Id: Idc7d60055949842023b85cd814801bb4b4b93022