deployment_server: Add Cassandra to autogenerated external svcs
The original config only queried the host IPs of the Cassandra nodes,
which are not the IPs that Cassandra is contacted on (more than one
Cassandra cluster may have a node running on a given machine).

So instead, query Cassandra instances via PQL, and massage the resulting
data into a dictionary that has lists of IPs by DC (eqiad, codfw, ...),
cluster name (ml_cache, cassandra-dev, ...), and instance (a, b, c...).

This results in a hash that has keys of the form `cassandra-dev_a_codfw`
that points to lists of IPs in that context. These names can then be
used in deployment charts (see change 1021895 for an example).

Bug: T360428
Change-Id: I64af34c6233561efeb351d299415840786feabaa