contint: set bullseye docker version just for host contint1002
During reimage we got:

E: Version '5:20.10.12~3-0~debian-bullseye' for 'docker-ce' was not found

Set the bullseye version for docker-ce just for this host,
without changing the global default yet.

[apt1002:~] $ sudo -E reprepro ls docker-ce
..
docker-ce | 5:20.10.18~3-0~debian-bullseye | bullseye-wikimedia | amd64

Try to get a succesful puppet run before the cookbook gives up on
detecting one to avoid a failed cookbook run and a limbo state.

Version string without the suffix because puppet code is:

$full_docker_version = "${docker_version}-${::lsbdistcodename}"

Bug: T334517
Change-Id: I7a9daf57df0578751f2bf49905fb67d09d6a27e5