Disable CustomVariables and CustomPiwikJs on new matomo server
We are in the process of deploying a new version of Matomo, but the
config.ini.php is currently shared by both the old and new instance.

We need to disable these two modules on the new instance. When we have
decommissioned the older instance, we will be able to remove the
references.

Hosts: matomo1002.eqiad.wmnet,matomo1003.eqiad.wmnet
Bug: T349397
Change-Id: Iefe69522d4647dfa9532f219b5a0c269bf17d490