Swith matomo/piwik to the new host
We are upgrading our Matomo instance by replacing the buster host with a
new host that runs bullseye and a more recent release of Matomo.

In order to effect the change we need to chage the destination host in
the trafficserver configuraiton and apply this to all cp hosts.

Hosts: O:cache::text
Bug: T351552
Bug: T349397
Change-Id: I0f5354b396fe242c53188f301eb09468e9cf727e