ci: avoid hardcoded IP in Hiera, lookup contint.wikimedia.org
Since the IP of the gearman_server here is always the IP of the
primary server and the primary server is always the one behind
the CNAME contint.wikimedia.org..

.. we can as well look that up on the fly and avoid having a
hardcoded IP and another thing we have to change during migrations.

Uses only v4 IP, so no change to before.

Bug: T334517
Change-Id: Ieabd362b6689a888811837478b0eb49a2e2a176f