Add the verbose flag to the geoipupdate command
We are concerned that the maxmind databases are not being properly
updated since a recent change to their file hosting configuration.

Specifically, the files that we have downloaded are not showing updated
timestamps for the last two days, although the systemd service running
geoipupdate claims success.

We should be able to remove this configuration once we have ascertained
whether or not there is a problem to resolve.

Bug: T358268
Change-Id: Ib8c2733905373625c54d1ff08cc6c0437bd08b67