lists: start a class for automating certain subscriptions
We have a set of VMs (stewards*) where we create lists of users
that should be subscribed to certain mailing lists for stewards.

There is an application, the onboarding-system, that is attempting
to automate the process.

The idea here is that we rsync lists of users from stewards* to lists*
and then run the "mailman-wrapper syncmembers" command.

This lets us keep list subscribers in sync with the generated files and
handle both subscribe and unsubscribe automatically, without having to
deal with API access.

https://docs.mailman3.org/projects/mailman/en/latest/src/mailman/commands/docs/syncmembers.html

On the lists* side this will be a directory that gets rsynced to with
rsync::quickdatacopy and a systemd timer/unit that will run the sync
command.

Just starting a class that creates the matching dir we have on the other
side so rsync::quickdatacopy can copy one to the other machine using
the same file system path.

Bug: T351202
Change-Id: I0a0269e6cd71454253646b5ad5a1617db1e740b0