hiera: buffer memory limit increase for cp4037
A very simple test to check if the UDP message drop is due to a
insufficient application buffer or not. The original memory limit is
just an hint from the official documentation and probably more tuning
will be useful in the future.

The buffer_memory_limit_bytes is now set to 700MB for this specific
host (cp4037)

Bug: T358109
Change-Id: I989ddb8e2dbc2fda237bba9a4880c4deee9797aa