wmcs VM backups: move all backups to one host, cloudbackup1004
These backup hosts need reimaging, and this is the safest
way to do that without running backupless during the change.
After this has had a few days to populate I'm going to reimage
cloudbackup1003, and then switch the backup load over there.

There enough room to run all backups on one host, although
it'll be a bit crowded.

Bug: T332400
Change-Id: I8e18fde81509cb8e44b366940d9f3fc2882e283d