Add server aliases to the cirrus/cfssl proxy config
We are about to start switching the elasticsearch::cirrus users to cfssl
instead of cergen. However, we noticed that the nginx configuration that
is generated when a host uses cfssl does not include the server aliases,
which may result in a failure to route the request to the correct
server.

Hosts: P:elasticsearch::cirrus
Bug: T360439
Change-Id: I924861a6ca827c24745a3213937872158a26e275