benthos:haproxy_cache: field renaming moved to grok pattern
Moved the field renaming (for required fields) from branch result_map
directly to the grok pattern, actually using the definitive name in
first place.

With type hints we also avoid explicitly calling the number() method
on fields of type int.

This should simplify and reduce the bloblang code in the branch
processor.

Bug: T363420
Change-Id: I8247fca2d2c267849c1bef5f27100b932aab39ae