postfix: add sasl auth header
For sasl auth we don't want Google to look at the first received header,
since it is from a client ip. For the exim config we hid the ip, here we
add an "Authenticated sender:" to the Received header. That may not be
sufficient, so we will verify in prod.

Bug: T317574
Change-Id: Icba385baab3e4fdf4ab6568817a8145ce9a1f5ef